+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+# [next] - unreleased
+- Add authentication - login page, protected views, logout
+
 # [0.1.0] - 2023-08-12
 - Initial release. Public URL shortener. Can list and add new URLs.